backend/requirements.txt
# Web framework
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# LLM & Embeddings
openai>=1.30.0
sentence-transformers>=3.0.0
FlagEmbedding>=1.2.0
transformers==4.44.2        # required for FlagEmbedding compat

# Vector DB
qdrant-client>=1.9.0

# Retrieval
rank-bm25>=0.2.2
underthesea>=6.8.4

# Utils
python-dotenv>=1.0.0
pydantic>=2.7.0
pyyaml>=6.0
httpx>=0.27.0

# Optional GPU acceleration
# torch>=2.3.0              # uncomment for GPU
